# Conflicts:
#	.gitignore
#	src/main/java/com/waregang/receiving_service/receiving_process/application/ReceivingProcessService.java
#	src/main/java/com/waregang/receiving_service/receiving_process/domain/model/WorkerReceivingSession.java

+++ src/main/java/com/waregang/receiving_service/receiving_process/infrastructure/dto/ReceivedContentDto.java
@@ -3,5 +3,5 @@ package com.waregang.receiving_service.receiving_process.infrastructure.dto;
 public record ReceivedContentDto(
         String containerLpn,
         String sku,
-        int quantity
+        Long quantity
 ) {}
